order some types
